Sync xrootd and add webdav. Change some logic around too.

+++ docker/CMSRucioClient/scripts/CMSRSE.py
@@ -3,9 +3,11 @@
 """
 Class definition for a CMS RSE. And script for updating RSE
 """
+import copy
 import logging
 import pprint
 import re
+import json
 
 from rucio.client.client import Client
 from rucio.common.exception import RSEProtocolNotSupported, RSENotFound
@@ -21,9 +23,12 @@ DOMAINS_BY_TYPE = {
     'temp': {'wan': {'read': 1, 'write': 1, 'third_party_copy': 1, 'delete': 1},
              'lan': {'read': 0, 'write': 0, 'delete': 0}},
 }
-RUCIO_PROTOS = ['SRMv2']
-IMPL_MAP = {'SRMv2': 'rucio.rse.protocols.gfalv2.Default'}
-DEFAULT_PORTS = {'gsiftp': 2811}
+RUCIO_PROTOS = ['SRMv2', 'XRootD', 'WebDAV']
+PROTO_WEIGHT = {'WebDAV':2, 'XRootD': 3, 'SRMv2': 1}
+IMPL_MAP = {'SRMv2': 'rucio.rse.protocols.gfalv2.Default',
+            'XRootD': 'rucio.rse.protocols.gfal.Default',
+            'WebDAV': 'rucio.rse.protocols.gfalv2.Default'}
+DEFAULT_PORTS = {'gsiftp': 2811, 'root': 1094, 'davs':1094}
 
 
 class CMSRSE:
@@ -46,7 +51,14 @@ class CMSRSE:
         self.settings = {}
         self.settings['deterministic'] = deterministic
         self.rucio_rse_type = json['type'].upper()
-        self.rse_name = json['rse']
+
+        # If we are building a _Test or _Temp instance add the special prefix
+        if cms_type=="test":
+            self.rse_name = json['rse']+"_Test"
+        elif cms_type=="temp":
+            self.rse_name = json['rse']+"_Temp"
+        else:
+            self.rse_name = json['rse']
 
         # pdb.set_trace()
 
@@ -151,9 +163,19 @@ class CMSRSE:
 
         if protocol_name not in RUCIO_PROTOS:
             return algorithm, proto
+        if proto_json['access'] != 'global-rw':
+            return algorithm, proto
 
-        domains = DOMAINS_BY_TYPE[self.cms_type]
+        domains = copy.deepcopy(DOMAINS_BY_TYPE[self.cms_type])
 
+        try:
+            for method, weight in domains['wan'].items():
+                if weight and protocol_name in PROTO_WEIGHT:
+                    domains['wan'][method] = PROTO_WEIGHT[protocol_name]
+        except KeyError:
+            pass  # We're trying to modify an unknown protocol somehow
+
+        #TODO: Make sure global-rw is set
         if proto_json.get('prefix', None):
             """
             The simple case where all we have is a prefix. This just triggers the identity algorithm 
@@ -184,16 +206,43 @@ class CMSRSE:
                     prefix = '/' + prefix_match.group(4)
                     extended_attributes = None
                 except AttributeError:
-                    print("Looking for minimal")
+                    try:
+                        print("Looking for prefix, no port and suffix")
+                        prefix_regex = re.compile(r'(\w+)://([a-zA-Z0-9\-\.]+)/(.*)')
+                        prefix_match = prefix_regex.match(proto_json['prefix'])
 
-                    prefix_regex = re.compile(r'(\w+)://([a-zA-Z0-9\-\.]+)/(.*)')
-                    prefix_match = prefix_regex.match(proto_json['prefix'])
+                        scheme = prefix_match.group(1)
+                        hostname = prefix_match.group(2)
+                        extended_attributes = None
+                        prefix = '/' + prefix_match.group(3)
+                        port = DEFAULT_PORTS[scheme]
+                    except AttributeError:
+                        try:
+                            print("Looking for prefix, port and no suffix")
+                            prefix_regex = re.compile(r'(\w+)://([a-zA-Z0-9\-\.]+):(\d+)')
+                            prefix_match = prefix_regex.match(proto_json['prefix'])
 
-                    scheme = prefix_match.group(1)
-                    hostname = prefix_match.group(2)
-                    extended_attributes = None
-                    prefix = '/' + prefix_match.group(3)
-                    port = DEFAULT_PORTS[scheme]
+                            scheme = prefix_match.group(1)
+                            hostname = prefix_match.group(2)
+                            extended_attributes = None
+                            prefix = '/'
+                            port = prefix_match.group(3)
+                        except AttributeError:
+                            print("Looking for prefix, no port and no suffix")
+                            prefix_regex = re.compile(r'(\w+)://([a-zA-Z0-9\-\.]+)')
+                            prefix_match = prefix_regex.match(proto_json['prefix'])
+                            scheme = prefix_match.group(1)
+                            hostname = prefix_match.group(2)
+                            extended_attributes = None
+                            prefix = '/'
+                            port = DEFAULT_PORTS[scheme]
+
+            # if we are building a _Test instance add the specia prefix
+            if self.cms_type == "test":
+                prefix = prefix +"/store/test/rucio/"
+
+            elif self.cms_type == "temp":
+                prefix = prefix +"/store/temp/"
 
             proto = {
                 'scheme': scheme,
@@ -213,7 +262,6 @@ class CMSRSE:
             tfc = []
 
             algorithm = 'cmstfc'
-            prefix_regex = re.compile(r'(\w+)://([a-zA-Z0-9\-\.]+):(\d+)(\/.*\=)(.*)')
             rules = proto_json['rules']
             for rule in rules:
                 # On first pass, fill in the basic information we need before pulling out the rules
@@ -221,6 +269,7 @@ class CMSRSE:
                 if rule.get('chain', None):
                     chains.add(rule['chain'])
                 try:
+                    prefix_regex = re.compile(r'(\w+)://([a-zA-Z0-9\-\.]+):(\d+)(\/.*\=)(.*)')
                     prefix_match = prefix_regex.match(rule['pfn'])
                     proto['scheme'] = prefix_match.group(1)
                     proto['hostname'] = prefix_match.group(2)
@@ -232,6 +281,7 @@ class CMSRSE:
                     proto['impl'] = IMPL_MAP[protocol_name]
                 except AttributeError:
                     prefix_regex = re.compile(r'(\w+)://([a-zA-Z0-9\-\.]+)(.*)')
+                    prefix_match = prefix_regex.match(rule['pfn'])
                     proto['scheme'] = prefix_match.group(1)
                     proto['hostname'] = prefix_match.group(2)
                     proto['port'] = DEFAULT_PORTS[proto['scheme']]
@@ -240,16 +290,12 @@ class CMSRSE:
                     proto['domains'] = domains
                     proto['impl'] = IMPL_MAP[protocol_name]
 
-                    print('Replace with actual no match error')
-                    raise
-
             # Now go through all the protocols including ones we were not interested in at first and get rules
             # Turn {"protocol": "SRMv2",
             #       "access": "global-rw",
             #         "rules": [ {"lfn": "/+(.*)", "pfn": "srm://SRM_URL=/$1", "chain": "pnfs"} ]
             #      },
             # into {u'path': u'(.*)', u'out': u'/pnfs/gridka.de/cms$1', u'proto': u'srmv2', 'chain': 'pnfs'}
-
             while chains - done_chains:
                 for test_proto in protos_json:  # Keep looking for what we need in all the protos
                     proto_name = test_proto['protocol'].lower()
@@ -263,7 +309,43 @@ class CMSRSE:
                             tfc.append(entry)
                         done_chains.add(proto_name)
 
-            proto['extended_attributes']['tfc'] = tfc
+
+            # If we are building a _Test or _Temp instance
+            if self.cms_type == "test" or self.cms_type == "temp":
+
+                # We need to find the rule that applies to the special prefix
+                # used for _Test(/store/test/rucio) or _Temp(/store/temp) and
+                # adpat it as a prefix
+                for rule in tfc:
+                    prefix_regex = re.compile(rule['path'])
+                    if self.cms_type =="test":
+                        prefix_match = prefix_regex.match("/store/test/rucio")
+                    else:
+                        prefix_match = prefix_regex.match("/store/temp")
+                    if prefix_match:
+                        match_rule = rule['path']
+                        g1 = prefix_match.group(1)
+                        out = rule['out']
+                        prefix = out.replace("$1", g1)
+                        break
+
+                # remove the part of the scheme://hostname...
+                prefix_regex1 = re.compile(r'(\w+)://([a-zA-Z0-9\-\.]+):(\d+)/(.*)')
+                prefix_regex2 = re.compile(r'(\w+)://([a-zA-Z0-9\-\.]+)/(.*)')
+                prefix_match1 = prefix_regex1.match(prefix)
+                prefix_match2 = prefix_regex2.match(prefix)
+                if prefix_match1:
+                    prefix = "/"+prefix_match1.group(4)
+                elif prefix_match2:
+                    prefix = "/"+prefix_match2.group(3)
+                else:
+                    logging.error("something went wrong when trying to calculate the special prefix")
+
+                proto['prefix']= prefix
+                proto['extended_attributes']= None
+
+            else:
+                proto['extended_attributes']['tfc'] = tfc
 
         pprint.pprint(proto)
         return algorithm, proto
@@ -276,26 +358,34 @@ class CMSRSE:
         except (RSEProtocolNotSupported, RSENotFound):
             current_protocols = []
 
-        protocol_unchanged = False
-
         for new_proto in self.protocols:
-
+            protocol_unchanged = False
             for existing_proto in current_protocols:
                 if existing_proto['scheme'] == new_proto['scheme']:
                     if new_proto == existing_proto:
                         protocol_unchanged = True
                     else:
-                        logging.info("Deleting definition which is not as expected: \nrucio=%s  \nexpected=%s",
-                                     str(existing_proto), str(new_proto))
-                        try:
-                            self.rcli.delete_protocols(rse=self.rse_name, scheme=new_proto['scheme'])
-                        except RSEProtocolNotSupported:
-                            logging.debug("Cannot remove protocol %s from %s", new_proto['scheme'], self.rse_name)
+                        if self.dry:
+                            logging.info("Deleting definition which is not as expected (Dry run, skipping)")
+                            logging.info(json.dumps(existing_proto, sort_keys=False, indent=4))
+                            logging.info("expected (Dry run, skipping)")
+                            logging.info(json.dumps(new_proto, sort_keys=False, indent=4))
+                        else:
+                            logging.info("Deleting definition which is not as expected:")
+                            logging.info(json.dumps(existing_proto, sort_keys=False, indent=4))
+                            logging.info("expected:")
+                            logging.info(json.dumps(new_proto, sort_keys=False, indent=4))
+                            try:
+                                self.rcli.delete_protocols(rse=self.rse_name, scheme=new_proto['scheme'])
+                            except RSEProtocolNotSupported:
+                                logging.debug("Cannot remove protocol %s from %s", new_proto['scheme'], self.rse_name)
 
-            if new_proto['scheme'] in ['srm', 'srmv2', 'gsiftp'] and not protocol_unchanged:
-                logging.info('Adding %s to %s', new_proto['scheme'], self.rse_name)
-                self.rcli.add_protocol(rse=self.rse_name, params=new_proto)
-
+            if new_proto['scheme'] in ['srm', 'srmv2', 'gsiftp', 'root', 'davs'] and not protocol_unchanged:
+                if self.dry:
+                    logging.info('Adding %s to %s (Dry run, skipping)', new_proto['scheme'], self.rse_name)
+                else:
+                    logging.info('Adding %s to %s', new_proto['scheme'], self.rse_name)
+                    self.rcli.add_protocol(rse=self.rse_name, params=new_proto)
         return
 
     def _create_rse(self):
